wip: extract completion wizard into nowPlayingCompletion.handler.ts
Work in progress -- paused mid-task due to token limit. See PR description for
what has been done and what remains.
